# vafund-backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API untuk aplikasi VaFund menggunakan Go Fiber dan Hyperledger Fabric",
        "title": "VaFund Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@vafund.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/withdrawals": {
            "get": {
                "description": "Get list of all withdrawals",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "withdrawals"
                ],
                "summary": "Get all withdrawals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawalsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawalsResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new withdrawal record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "withdrawals"
                ],
                "summary": "Create withdrawal",
                "parameters": [
                    {
                        "description": "Withdrawal data",
                        "name": "withdrawal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawalResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawalResponse"
                        }
                    }
                }
            }
        },
        "/api/withdrawals/event/{event_code}": {
            "get": {
                "description": "Get list of withdrawals for a specific event",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "withdrawals"
                ],
                "summary": "Get withdrawals by event code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Code",
                        "name": "event_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawalsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawalsResponse"
                        }
                    }
                }
            }
        },
        "/api/withdrawals/event/{event_code}/total": {
            "get": {
                "description": "Get total amount of withdrawals for a specific event",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "withdrawals"
                ],
                "summary": "Get total withdrawal amount by event code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Code",
                        "name": "event_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/withdrawals/total": {
            "get": {
                "description": "Get total amount of all withdrawals",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "withdrawals"
                ],
                "summary": "Get total withdrawal amount",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/withdrawals/{id}": {
            "get": {
                "description": "Get a specific withdrawal by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "withdrawals"
                ],
                "summary": "Get withdrawal by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Withdrawal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawalResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawalResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawalResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Withdrawal": {
            "type": "object",
            "required": [
                "amount",
                "date_time",
                "event_code",
                "withdraw_by"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "minimum": 0.01
                },
                "date_time": {
                    "type": "string"
                },
                "event_code": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "tx_id": {
                    "type": "string"
                },
                "withdraw_by": {
                    "type": "string"
                }
            }
        },
        "models.WithdrawalRequest": {
            "type": "object",
            "required": [
                "amount",
                "event_code",
                "withdraw_by"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "minimum": 0.01
                },
                "event_code": {
                    "type": "string"
                },
                "withdraw_by": {
                    "type": "string"
                }
            }
        },
        "models.WithdrawalResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Withdrawal"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.WithdrawalsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Withdrawal"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}